pyproject.toml
[project]
name = "gopro"
version = "0.1.0"
description = "GoPro Hero 13 roast sentinel — vision monitoring for coffee roasting"
requires-python = ">=3.12"
dependencies = [
    "open-gopro>=0.16.0",
    "anthropic>=0.79.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "websockets>=16.0",
]
